requirements.txt
# gRPC
grpcio>=1.62.0
grpcio-tools>=1.62.0

# LangChain and LangGraph
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
langgraph>=0.0.20

# LLM providers (add as needed)
# openai>=1.12.0
# anthropic>=0.18.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Logging
structlog>=24.1.0

# Optional: For semantic routing
# sentence-transformers>=2.3.0
# numpy>=1.26.0

